Use sqlc (#74)
* introduces sqlc, schema, queries

* all table queries in sqlc

* removed gorm in favor of sqlc... WIP adjusting code

* wip

* WIP 2

* WIP 3

* WIP 4

* fixing Nullable values Valid field

* first working run

* enable foreign_keys on sqlite

* fixing tests 1

* fixed tests

* docker build with sqlc

* fixed Dockerfile

* helpful local makefile to build/clean

* fixed bug on log will message

+++ jwt/jwt.go
@@ -7,7 +7,7 @@ import (
 	jwt "github.com/golang-jwt/jwt/v5"
 )
 
-func CreateToken(signingKey []byte, issuer string, hoursDuration int, userId uint) (string, error) {
+func CreateToken(signingKey []byte, issuer string, hoursDuration int, userId int64) (string, error) {
 
 	claims := TagyouClaims{
 		userId,
